refactor: extract MomentsSection, move HowSection CTA to header
- Moments table extracted from GetSection into its own MomentsSection
  component with proper section py spacing
- HowSection CTA "Get to a personalised demo" moved to header area,
  right-aligned as a primary button
- Tests split accordingly (6 moments tests → MomentsSection.test.tsx)
