+++ src/components/Landing/__tests__/GetSection.test.tsx
@@ -51,63 +51,4 @@ describe("GetSection", () => {
     });
   });
 
-  describe("moments table", () => {
-    it("renders the moments table h2", () => {
-      render(<GetSection />);
-      expect(
-        screen.getByRole("heading", {
-          level: 2,
-          name: "Six months in, you own something.",
-        })
-      ).toBeInTheDocument();
-    });
-
-    it("renders the moments table lede", () => {
-      render(<GetSection />);
-      expect(
-        screen.getByText(/Other automation tools leave you with rigid workflows which decay/)
-      ).toBeInTheDocument();
-    });
-
-    it("has 3 column headers", () => {
-      render(<GetSection />);
-      expect(
-        screen.getByRole("columnheader", { name: "What you have" })
-      ).toBeInTheDocument();
-      expect(
-        screen.getByRole("columnheader", { name: "With automation" })
-      ).toBeInTheDocument();
-      expect(
-        screen.getByRole("columnheader", { name: "With Colex" })
-      ).toBeInTheDocument();
-    });
-
-    it("has 6 data rows with correct first cells", () => {
-      render(<GetSection />);
-      const firstCells = [
-        "Your company’s standards, written down",
-        "A record of every past judgement",
-        "Standards that get sharper with use",
-        "An answer for the auditor",
-        "A process anyone can pick up",
-        "Books that stay honest after the fact",
-      ];
-      for (const cell of firstCells) {
-        expect(screen.getByText(cell)).toBeInTheDocument();
-      }
-    });
-
-    it("uses semantic <table> markup", () => {
-      render(<GetSection />);
-      expect(screen.getByRole("table")).toBeInTheDocument();
-    });
-
-    it("table scrolls on mobile (overflow-x: auto)", () => {
-      render(<GetSection />);
-      const table = screen.getByRole("table");
-      const scrollContainer = table.closest("[data-testid='table-scroll']");
-      expect(scrollContainer).toBeInTheDocument();
-      expect(scrollContainer).toHaveStyle({ overflowX: "auto" });
-    });
-  });
 });
